set a structure for the parsing according to BNF notation
