+++ src/1-Input-Management/lexer/expand_vars.c
@@ -6,7 +6,7 @@
 /*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/24 12:58:49 by anshovah          #+#    #+#             */
-/*   Updated: 2023/10/01 11:27:59 by anshovah         ###   ########.fr       */
+/*   Updated: 2023/10/02 15:39:31 by anshovah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
