+++ src/1-Input-Management/parser/parser.c
@@ -1,0 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   parser.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/13 16:16:49 by astein            #+#    #+#             */
+/*   Updated: 2023/10/02 18:32:44 by anshovah         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void	parse(t_minibox *minibox)
+{
+    minibox->tmp_token = minibox->tokens;
+    command(minibox);
+    //TODO:
+}
