+++ src/1-Input-Management/parser/redir.c
@@ -1,0 +1,31 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   redir.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: anshovah <anshovah@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/10/02 17:50:45 by anshovah          #+#    #+#             */
+/*   Updated: 2023/10/02 18:36:18 by anshovah         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+# include "minishell.h"
+
+static t_tree *redir_case1(t_minibox *minibox);
+static t_tree *redir_case2(t_minibox *minibox);
+
+t_tree *redir_main(t_minibox *minibox)
+{
+}
+
+// <redir in>
+static t_tree *redir_case1(t_minibox *minibox)
+{
+
+}
+// <redir out>
+static t_tree *redir_case2(t_minibox *minibox)
+{
+    
+}
